refactor(#45): remove element state from DragControlContext
