+++ src/Node.tsx
@@ -1,12 +1,7 @@
 import React, { useEffect, useRef, useContext, useCallback } from "react";
 import { getEmptyImage } from "react-dnd-html5-backend";
 import { Container } from "./Container";
-import {
-  useDragNode,
-  useDropNode,
-  useDragSourceElement,
-  useDragControl,
-} from "./hooks";
+import { useDragNode, useDropNode, useDragControl } from "./hooks";
 import { NodeModel, RenderParams } from "./types";
 import { TreeContext } from "./Tree";
 
@@ -43,7 +38,6 @@ export const Node: React.FC<Props> = (props) => {
     }
   }, []);
 
-  useDragSourceElement(ref);
   useDragControl(ref);
 
   const handleToggle = useCallback(() => {
