+++ src/hooks/index.ts
@@ -1,7 +1,6 @@
 export { useDragControl } from "./useDragControl";
 export { useDragNode } from "./useDragNode";
 export { useDragOver } from "./useDragOver";
-export { useDragSourceElement } from "./useDragSourceElement";
 export { useDropContainer } from "./useDropContainer";
 export { useDropNode } from "./useDropNode";
 export { useOpenIdsHelper } from "./useOpenIdsHelper";
